Wire full loan lifecycle end-to-end with BFF and demo auth
Implement the loan_application aggregate (19-state machine, SoD, audit
trail), integer-paise LoanMath, Flyway V3–V9, and 114+ backend tests.
Add Next.js BFF with separate borrower/staff sessions, live staff
pipelines, applicant KYC/doc review, and wire the designed UI to real
APIs. Document current state and go-live backlog in CLAUDE.md, handoff,
dfd, and FUTURE.md.

+++ backend/navix-collections/src/main/java/com/navix/collections/entity/Settlement.java
@@ -10,7 +10,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -35,8 +34,9 @@ public class Settlement {
     @Column(name = "collection_case_id", nullable = false)
     private UUID collectionCaseId;
 
+    /** Agreed settlement amount, in paise. */
     @Column(name = "settlement_amount", nullable = false)
-    private BigDecimal settlementAmount;
+    private Long settlementAmount;
 
     @Column(name = "proposed_by", nullable = false)
     private UUID proposedBy;
